+++ src/kernel/arch/aarch64/include/cpu/cpu.hpp
@@ -23,8 +23,8 @@
 #include <typeinfo>
 
 #include "kernel_log.hpp"
+#include "regs.hpp"
 #include "sk_cstdio"
 #include "sk_iostream"
-#include "sr.hpp"
 
 #endif  // SIMPLEKERNEL_SRC_KERNEL_ARCH_AARCH64_INCLUDE_CPU_CPU_HPP_
